Fixed numpy error in sunpy/image/rescale.py

+++ sunpy/image/rescale.py
@@ -225,8 +225,8 @@ def reshape_image_to_4d_superpixel(img, dimensions, offset):
 
     # Reshape up to a higher dimensional array which is useful for higher
     # level operations
-    return (img[offset[0]:offset[0] + na*dimensions[0],
-                offset[1]:offset[1] + nb*dimensions[1]]).reshape(na, dimensions[0], nb, dimensions[1])
+    return (img[int(offset[0]):int(offset[0] + na*dimensions[0]),
+                int(offset[1]):int(offset[1] + nb*dimensions[1])]).reshape(na, dimensions[0], nb, dimensions[1])
 
 
 class UnrecognizedInterpolationMethod(ValueError):
